bs-dkg test moved
